properly handle duplicate events for an fd in poll notify functions
This addresses the scenario where check_fdesc() (invoked from epoll_wait) and
wait_notify() (invoked from notify_dispatch()) could both invoke
epoll_wait_notify(), poll_notify() or select_notify() for a given fd. This is
a reasonable occurance as a call to epoll_wait(), poll() or select() could
pick up a coincident event state change prior to completion of the ensuing
notify_dispatch(). In the case of epoll_wait(), a race could lead to duplicate
epoll_events being reported to userspace. For poll() and select(), an
incorrect count of events could be reported. This flaw was detected in low
frequency in the signalfd portion of the signal runtime test. In this case, a
duplicate epoll events entry led to a superfluous call to read(2), which would
block indefinitely.

epoll_wait_notify() will now scan the reported events array to look for a
matching data field and, if found, update the associated events field with the
union of the two reported events. (It is necessary to use the union of events
rather than supplant the events in the array, otherwise edge-triggered events
could be omitted.) poll_notify() and select_notify() have been changed to only
increment the event count for events not previously reported.