blockq_check_timeout(): protect against race between action invocation and inserting waiter
The blockq previously had a known limitation whereby a wakeup occurring
between initial invocation of the blockq action and insertion into the waiting
list could go undetected. While the futex code had a specific workaround for
this, this race has since led to missed wakeups in other areas, calling for a
more general solution.

With this change, blockq_wake_one(), under blockq lock, sets the wake flag if
a waiting thread could not be dequeued. blockq_check_timeout() firsts clears
the flag prior to invoking the action, then checks (under blockq lock) if the
flag was set by a wake within this interval. If it was, a wakeup may (but not
necessarily) have been missed, and blockq_wake_one() is called after releasing
the locks.

This solution may lead to a spurious wakeup in the case that the state change
indicated by the wake was in fact detected by the blockq action. This would
seem to affect only the futex code, however such spurious futex wakeups (a
FUTEX_WAIT returning 0, not triggered by a FUTEX_WAKE) are explictly
allowed. (See the FUTEX WAIT entry under the RETURN VALUE section of the
futex(2) manpage.) Following this pattern (which appears to be common amongst
pthreads and posix interfaces), all blockq actions should be written to assume
that spurious invocations are possible.

The aforementioned workaround in the futex code (spinning on non-zero waiters
for a wake) has been removed.