Add print_tree command to gdb to print rbtree
This change adds a new python gdb command, print_tree, to print the pointers of
rbnodes of an rbtree in order. If a second argument, containing type, is passed
the script will print the pointers to the containing struct instead of the
rbnode. Also added are helper macros offsetof() and container_of() which can
be used in gdb expressions.

Also added is a new array called 'processes' containing the pointers to the two
process structs in order to facilitate easier process debugging.