UEFI loader: exit boot services if halt() is called
Instead of borrowing the halt() function implementation from the
kernel code, halt() is now implemented in the UEFI loader so that
boot services are exited when there is a fatal error. With this
change, a halt() call, instead of freezing the machine, allows the
UEFI firmware to run an interactive shell, which can be useful for
debugging.