Aarch64: implement kernel booting via UEFI
This commit adds a UEFI loader for aarch64 and enhances the aarch64
kernel with ACPI support.
When the ENABLE_UEFI option is specified in the Makefile command
line (as in `make PLATFORM=virt ENABLE_UEFI=1`), the resulting
image contains a UEFI partition with the loader, which allows the
image to boot on UEFI-based aarch64 machines.
The UEFI loader uses a new functionality of the ELF code to load
the program sections of the kernel to their physical addresses
(the UEFI firmware sets up the MMU with identity mapping), and
then jumps to the physical address of the kernel entry point.
The physical memory map and the root pointer of the ACPI tables
are passed by the loader to the kernel, which uses this information
when creating the physical memory id heap and to parse the ACPI
tables.
The base adresses of the PCI ECAM and GIC memory-mapped address
spaces are probed via ACPI, and the existing hardcoded address
values are only used if ACPI probing fails. The serial port
peripheral is also probed via ACPI.
The GIC driver has been enhanced to support MSI interrupts in
GICv3, which use the ITS (Interrupt Translation Service).

These changes allow running Nanos on AWS Graviton instances, which
are aarch64-based machines with NVMe disks and ENA network
adapters.