netconsole_write(): check if lwIP mutex is acquired
Messages sento to the console may come from lwIP callbacks (e.g.
when an IP address is acquired via DHCP). In these cases, the lwIP
lock is already held by the current context, and thus trying to
lock it leads to a "lock already held" error.
This change defines a `mutex_is_acquired()` macro that returns
whether a given mutex is held by the current context; this macro is
used in netconsole_write() in order to lock the lwIP mutex only if
it is not already held by the current context.