use an MCS lock for mutex spinwaiters
This updates the two-phase mutex to use a cancelable MCS (Mellor-Crummey and
Scott) lock, which allows only one spin waiter to contend for the mutex at a
time. As before, a mutex lock operation will spin for only m->spin_iterations
before reverting to suspending the context. Before suspending, the cpuinfo is
removed from the MCS list, and the corresponding context is added to the
mutex's queue of waiters. (The waiters list is now a standard doubly-linked
list, protected by a spinlock within the mutex.) If the MCS lock was acquired
while spinning, the mutex itself is then taken and the MCS lock is released,
passing ownership to the next queued spinner, if there is one.

On an unlock, the mutex is released by zeroing m->turn. This may allow the
current holder of the MCS lock, which is spinning while trying to acquire
m->turn, to proceed with taking the mutex. If there is a (suspended) context
waiting on the mutex, it is treated with equal priority to the MCS owner and
is scheduled to be resumed. The resumed context then spins to acquire the
mutex.

While MCS is relatively straightforward to implement, the ability to exit the
MCS queue - critical to prevent cores from spinning indefinitely - is a
somewhat delicate process that has been adapted from the optimistic spin queue
implementation in Linux by Peter Zijlstra, et al.