Reorder mkfs debug_exit argument before image argument
The BSD/OS X implementation of getopt stops processing at the first non-option
argument. The mkfs command had the debug_exit option specified after the image
and so that option was not getting processed by getopt, causing test failures on OS X
not to be registered. This change moves the debug_exit argument with the others
so that it will be processed.