Storage drivers: add scatter-gather I/O capability
This commit changes the interface between storage drivers and the
rest of the kernel so that all storage requests are made via a
single storage_req_handler closure, whose function takes a
storage_req structure containing the parameters for the request
being issued. Among these parameters there is an opcode indicating
the request type; the STORAGE_OP_READSG and STORAGE_OP_WRITESG
opcodes are used for scatter-gather I/O.
This commit implements scatter-gather I/O in the virtIO block and
virtIO SCSI drivers, while the remaining storage drivers use a
generic helper function that maps a given request to the existing
read and write block_io closures.

With this change, average read speed values for a bulk read test
program changed from 113 to 283 MB/s. Test conditions: reading 4
files large 128 MB each, with read requests of 128 KB each, running
on qemu (virtio-scsi disk device), with file contents not cached in
the guest RAM, but with the relevant parts of the img file cached
in the host RAM.