tracelog feature
This change introduces a trace logging feature into the Nanos kernel.
Debugging prints that were previously directed to the console may now be
buffered for later retrieval using the tprintf() and vtprintf() functions.
The tracelog was designed to minimize the cost of adding trace entries, with
the ultimate aim of minimizing impact on execution time of the code being
analyzed. Each cpu has a dedicated trace buffer, avoiding the need to take a
lock when adding trace data. These per-cpu buffers are merged together (and
replaced with fresh buffers) by a background task that is scheduled when
buffer storage runs low or a request to download trace data is made (via the
tracelog HTTP interface on port 9091). Trace entries may be filtered based on
a "tag" symbol as well as optional attributes specified in the trace (e.g. a
thread id for unix thread traces).

The tracelog is only built when "TRACE=tracelog" is specified on the make
commandline. Tracing output from various parts of the kernel may be selected
individually using a comma-separated DEBUG variable, with DEBUG=all indicating
that all traces should be enabled. If DEBUG is specified without building
support for the tracelog, trace output is directed to the console.

For more details on tracelog usage, see doc/tracelog.txt