getsockopt/setsockopt: fix handling of TCP_NODELAY socket option
If the socket on which getsockopt/setsockopt is called is in
listening state, it is not possible to get or set TCP options on
the lwIP PCB structure, because the structure associated to a
listening socket is a struct tcp_pcb_listen, not a struct tcp_pcb.
A struct tcp_pcb_listen does not have a TCP flags field, thus
trying to access this field leads to a potential corruption of
other fields in struct tcp_pcb_listen.
This commit fixes the above issue by adding a tcpflags_t field to
the netsock structure and accessing this field in
getsockopt/setsockopt if there is no struct tcp_pcb to be accessed.
To match the Linux implementation, any socket created from a
listening socket inherits its TCP options.