id heap: implement next-fit search based on allocation size
Using a first-fit allocation policy in the id heap does not scale
well with heap size, because after a substantial amount of
allocations, for each subsequent allocation a large number of heap
locations needs to be scanned (especially for small-sized
allocations) before a free location is found.
Yet, a simple next-fit allocation policy leads to excessive memory
fragmentation when allocations of different sizes are interleaved,
because a large gap is formed when a large allocation follows a
small allocation (to satisfy the id heap alignment property).
This commit implements a more advanced next-fit allocation policy
where different next-fit locations are stored for different
allocation sizes: this prevents a large allocation from affecting
the next-fit pointer for smaller allocations and allows them to
fill the gaps left by large allocations.
In id_alloc_from_range(), a pre-existing issue in the calculation
of the `max_start` value (now renamed as `margin`) is being fixed.