Add preallocated versions of objcache and table
To prepare for lock profiling, versions of the objcache and table are needed
that do not require allocation from the shared heap during runtime. This
change adds a way to create these data structures which allocate memory
for a specified number of entries at initialization. The objcache can be
specified to allow additional allocations once the preallocated structures
are exhausted, if desired. The preallocated table uses a preallocated
objcache to provide table entries, but does not allow additional entry
allocations or table resizing.