virtio SCSI: add parsing of device identification VPD page
Google Cloud Platform allows a device name to be associated to a
volume when attaching the volume to an instance; the device name is
then reflected in the SCSI device identifier reported by the
attached volume. This change implements retrieval and parsing of
the device identification Vital Product Data page (where a
descriptor containing the device identifier is found) in order to
detect the volume attachment point.