socket_write_udp(): fail on sending to an unconnected socket
The netsock test was previously failing under Linux, with the write() in
netsock_test_basic() failing with the errno EDESTADDRREQ (89). This was a
result of a missing connect() on the datagram socket, a condition that was
silently being ignored by Nanos. socket_write_udp() has been updated to report
this error when attempting to send to an unconnected udp socket, and a test
for this condition, as well as the required call to connect(), have been added
to the netsock test.