add support in netlink for RTM_GETROUTE requests
This work allows IP routing table information to be extracted from the kernel
using RTM_GETROUTE netlink messages. As with the remainder of netlink code,
only IPv4 is supported at the moment. This support is sufficient for
retrieving the gateway address for each interface. The netlink runtime test
has been expanded to test this functionality as well as optionally dump
details from rtmsg responses received from the kernel.