implement SIOCGIFNAME socket ioctl, handle netdevice ioctls for any socket type (#1700)
This change implements the SIOCGIFNAME ioctl for getting an interface name by
index. It also moves the various SIOC?* netdevice-related ioctls from
netsock_ioctl() to socket_ioctl(), as they are valid for any type of socket,
regardless of family or type (as stated in netdevice(7)). These changes are
required for if_indextoname(3) to work properly.