unix_fault_handler: deliver SIGSEGV on user faults from syscall context
This addresses a common problem whereby the kernel can panic on a user page
fault that occurs within a syscall context. Such faults are usually caused by
the user program supplying an invalid buffer address, and as such should be
treated as segmentation violations. This handles such cases by marking the
relevant thread's current syscall as having been "abandoned," thereafter
delivering SIGSEGV to the violating process - just as if the fault had
occurred from within user mode. The abandoned syscall is "cleaned up" in
thread_return() before returning control to the program (albeit not
completely; see the comment in that function).

This also addresses a few cosmetic issues with output that occurs after
terminal signal delivery. To aid the debugging of such violations, frames of
the the offending thread and syscall_context (if applicable) are dumped.

The signal test has been amended with checks for SIGSEGV produced by zero page
and kernel memory accesses from userspace, as well as by a fault on a bad user
address from within a syscall context.