report faulting PC for SIGILL, SIGTRAP and SIGFPE signals
unix_fault_handler() was incorrectly reporting the "fault address" from the
thread frame as the siginfo si_addr for SIGILL, SIGTRAP and SIGFPE
signals. This fix instead records the PC of the faulting instruction as
si_addr for these signals, and also replaces the ambiguous
frame_return_address() function with new definitions of frame_fault_pc() for
each architecture. To check the function of frame_fault_pc(), the SIGILL test
has been expanded to verify that the memory referenced by si_addr points
precisely to the illegal instruction used to induce the signal.