aarch64 GIC: fix selection of distributor/redistributor registers
In the GICv3, the redistributor registers must be used (instead of
the distributor registers) to configure all interrupts that are
private to each CPU (i.e. the SGIs and PPIs).
The gic_set_int_target() function is being removed because it
cannot be used to configure the target CPU for SGIs and PPIs (the
relevant distributor registers are read-only, and there are no
corresponding registers in the redistributor).