aarch64: add SMP support
This change adds support for multiple CPUs in the ARM port.
The assembly code and the code in interrupt.c, gic.c and page.c has
been refactored so that separate functions are provided for code
that needs to be executed by each CPU.
The GICv3 driver now properly supports accessing the redistributor
interface from CPUs other than the boot CPU.
In the common kernel initialization code, the linear backed heap is
being passed (instead of the page backed heap) to
init_kernel_contexts() in order to avoid mapping and unmapping
operations (and thus TLB flushes) when bringing up secondary CPUs
(e.g. when calling init_cpuinfo()), because the TLB flush code
needs to access the cpuinfo structure of the current CPU, which may
not be available during CPU bringup.

Closes #1397