select(2): allow blocking indefinitely with zero nfds (#1708)
Our select implementation incorrectly assumed that a call with zero nfds
specified and an indefinite timeout should return immediately. Such a call may
legitimately be used to wait for signal reception, though, so
select_internal() has been changed here to enter an indefinite, interruptible
block whenever a NULL timeout is specified, regardless of the value of
nfds. The signal test now checks that such a call to select behaves in the
same manner as pause(2).