vdso data strucure: put arch-specific fields in arch_vdso_dat
This change replaces the x86-specific `platform_has_rdtscp` field
in the vdso data structure with a struct arch_vdso_dat, which
contains arch-specific data. This allows minimizing the struct size
based on the fields needed on each architecture.
In addition, the 'packed' attribute is removed from the vdso data
structure, because it is not necessary and causes accesses to most
64-bit fields to be unaligned, which could negatively affect
performance.