bitmap_alloc_internal(): optimize allocations with 64-bit stride
Allocations in the [33, 63] bit range are aligned to 64 bits and
handled with a 64-bit stride; therefore, the bitmap can be searched
using the same code as for allocations >= 64 bits (which is more
efficient than the code for smaller allocations). This change uses
the stride value (instead of the allocation size) to determine the
code path to search the bitmap; this allows allocations in the
[33, 63] bit range to be handled more efficiently, and prevents a
"shift out of bounds" error generated by UBSan when executing the
`mask <<= stride` instruction.