io-uring: ensure natural alignment for completion queue entries
When an io-uring context is set up, the location of the array of
completion queue entries depends on the number of submission queue
entries requested by userspace. This change prevents unaligned
accesses to the 64-bit user_data field in a completion queue entry
structure when userspace requests an odd number of submission queue
entries.