getdents/getdents64: align dirent structures to 8 bytes
This change adds appropriate padding to the dirent records
generated by getdents and getdents64, so that unaligned accesses to
multi-byte fields are avoided (at least when the linux_dirent
pointer supplied by userspace is aligned to 8 bytes).
The amount of padding is minimized based on the length of the name
of a given dirent (i.e. ensuring that there is at least one null
byte after the name).
In addition, this change fixes setting the d_type field in struct
linux_dirent64.