Add support for UBSan (undefined behavior sanitizer)
This change allows specifying an `UBSAN` Makefile variable that
enables instrumenting the compiler-generated code with instructions
to catch bugs that can result in undefined behavior. When UBSan
detects an error, it generates a software trap (on x86-64, an
"invalid opcode (UD2)" exception; on aarch64, a "brk" exception
that results in an ESR value of 0xf20003e8; on risc-v, a
breakpoint exception), which triggers a crash dump in the kernel.
UBSan checks for unaligned accesses and null pointers are being
disabled because the kernel contains legitimate uses of unaligned
accesses (mainly in third-party code such as acpica and lwIP) and
null pointers (in the x86-64 code for secondary CPU bringup).
`UBSAN=1` is being added to the make command line in the
build-pc-memdebug CircleCI job (now renamed to build-pc-debug) so
that UBSan is enabled as part of the CI checks.