Sockets: deferred TCP processing for reads and writes
This change defers several operations with the highest contention on the lwip
lock to queues that can be processed asynchronously from their originating
source so that TCP socket read and write syscalls can be executed without
needing to grab the lwip lock, thus improvement throughput and scaling
for SMP VMs. These improvements only apply to TCP sockets.

A singleton service thunk has been created to handle any deferred processing
for multiple kinds of operations.

The network driver receives packets by calling the input function on the
netif with the lwip lock held. This operation is deferred by including a closure
on the driver's pbuf struct to call input on the pbuf from the service thunk.

Transmit pbufs are freed by the network driver, requiring the lwip lock. The
internal lwip protection has been enabled with SYS_LIGHTWEIGHT_PROT and
spinlock implementation of sys_arch_[un]protect is provided so that pbuf
free no longer requires holding the lwip lock.

In lieu of the lwip lock, a spinlock has been added to the netsock structure
to protect the fields of the netsock struct. While reading received pbufs
does not require the lwip lock, the read syscall for a TCP socket must
call lwip's tcp_recved for any read bytes which does require the lock.
This operation is deferred by tracking the number of read bytes on the
netsock and queueing a thunk to call tcp_recved from the service thunk,
which resets the read byte count.

Writes to a TCP socket have been modified to copy the user buffer to a new
kernel buffer, queued on the netsock, up to an arbitrary limit. These buffers
are processed in the service thunk where SGs are used to track how much
data has been sent by lwip. When lwip runs the callback after the data
has been acknowledged, the SG is consumed and the kernel buffers are
freed.