riscv64 serial_set_devbase(): fix pointer overflow error from UBSan
The undefined behavior sanitizer complains when a pointer
arithmetic operation causes a pointer value (interpreted as a
signed number) to change sign, as happens with the UART0_BASE
variable in the RISC-V serial driver when it is converted from a
physical address to a virtual address.
To avoid the above issue, UART0_BASE is being cast to an unsigned
number before doing the arithmetic operation.