ACPI: add support for entering S3 sleeping state (VM suspend)
This change implements a handler for the ACPI sleep button event,
which is triggered in a GCP instance during a VM suspend operation.
This handler communicates with the ACPI firmware to put the system
in the ACPI S3 state, at which point the VM is suspended; then,
when a VM resume operation is requested, guest execution resumes
from where it was suspended (at the call to AcpiEnterSleepState()),
after which the sleep button handler indicates to the ACPI firmware
that wakeup has completed and the VM resumes normal operation.