pagecache: track dirty pages on pagecache_node
If all references to an fsfile are removed, the pagecache_node for the file
also gets deallocated. If there were dirty pages or outstanding i/o, a crash
would occur when the i/o was completed as the pagecache_page had been
deallocated when the node was deallocated. This change adds a refcount
to the node to track dirty pages and only deallocates the node once all
dirty or writing pages have been flushed. Because the deallocation may be
called when the pagecache state lock is already locked as well as not locked,
the deallocation, in the kernel the deallocation must be deferred to the bhqueue
to avoid state lock reentrance.

An unrelated fix is also made to fetch the next pagecache_page in
pagecache_write_sg_finish before potentially deallocating the page.