Add core dump functionality for crashing user programs
This change adds support for dumping an elf core file of a crashing user
program to "/core" on the filesystem for later retrieval. The dump is only
created if the root tuple contains an entry for "coredumplimit" greater than 0.
The limit determines how much of the core will be written before stopping.
The core file contains process status, signal information, aux vector contents,
thread registers, and a dump of all user vmaps that are not file backed.
Core dumps are available for all architectures.
The mkfs program has been modified to check for coredumplimit and make
the image size at least as large as the coredumplimit.

Core dumps may be retrieved by mounting the volume with tfs-fuse, or copying
out with ops.