Copy socket flags from listener at syscall accept
A socket generated by an accept syscall from a listening socket should inherit
the tcp flags at the time of the accept syscall, not when the connection is
initiated at the network stack. This way, any changes to the listener's tcp flags
between the initiation of the connection and the accept syscall are appropriately
reflected in the child.