TFS log_flush(): avoid invoking flush completion directly
If log_flush() is called when the TFS log is not dirty, it should
invoke the flush completion immediately, but when running in kernel
code the completion should be applied asynchronously, because
log_flush() is called with the TFS lock held, and the completion
may need to acquire the TFS lock (e.g. log_flush_completed in tfs.c
calls pagecache_sync_volume(), which can execute TFS storage write
operations).