join adjacent and similar vmaps, add vmap debugs, extend mmap test
The behavior of allocate_vmap() is changed here to merge adjacent vmaps when
possible, thus reducing entries and lookup times. It also simplifies logic for
operations like mremap, which no longer would need to handle multiple,
adjacent vmaps with equivalent properties when considering a contiguous range
of mapped space. In addition, a VMAP_DEBUG option and relevant debugs have
been added for tracing changes to process vmaps. allocate_vmap() is now the
external interface for allocating user address space, and as such takes the
vmap lock before calling allocate_vmap_locked().

A test has been included to induce splitting and merging of vmaps. While the
test cannot observe that the correct transformations have been made on process
vmaps, such transformations may be manually observed by building with
VMAP_DEBUG. Other additions to the mmap test include a test for hints (address
specified for non-fixed mmap) as well as making the zeropage and exec
protection tests optional - for testing the lowering of mmap_min_addr to 0 and
running without exec protection enabled (or running on Linux). With these
changes, the mmap test may now be run on Linux (with the exec option) as a
parity check.