ELF dynamic linker: fix relocations in absence of RELA dynamic tag
A klib binary file may require relocations to be applied even if
its ELF dynamic section does not have a RELA tag. In this case,
the pointer to the relocation table is given by the JMPREL tag.
When both RELA and JMPREL tags are present, their pointer values
are the same, thus any one of these tags can be used to retrieve
the relocation table.
This change adds support for the JMPREL tag, which allows loading
klib binary files without a RELA tag.