Klibs: add support for in-kernel tests
In-kernel tests can be used to check the correct behavior of kernel
code that cannot be easily exercised via syscalls. In order to
avoid bloating the main kernel binary, in-kernel tests can be
implemented as klibs.
This change adds to the klib source folder a test subfolder that
will contain the source of klibs for in-kernel tests. In image
files, klib binaries are meant to be installed in a test subfolder
of the klib root directory; if the root tuple contains a
`klib_test` attribute, the kernel loads all klibs found in the test
subfolder, which will then execute the in-kernel tests.
The existing "test" klib (which checks the correct behavior of the
klib loading code) is being moved to the test folder, and the
"klibs" runtime test application is being renamed as "ktest"
(because it will run different in-kernel tests).
Klibs containing in-kernel tests are not included in tarballs
created by CircleCI automated builds; CircleCI jobs have been
modified in order to enable extended globbing in the shell options,
which is disabled by default in CircleCI containers and is needed
in order to exclude the test subfolder from the klib binary folder
when creating a tarball.

Closes #937.