mmap(2): validate flags parameter, implement MAP_POPULATE
The mmap(2) implementation was insufficiently checking the flags parameter,
selecting for implemented flags and ignoring all others. This change
explicitly checks for known options and validates that unknown options are not
blindly accepted. Support for MAP_POPULATE has been added, and the mmap test
now checks the handling of unaligned hint and fixed addresses.