mremap(2) fixes and improvements
The mremap(2) implementation contained a flaw whereby the remapping of a
portion of a vmap would lead to the entire vmap being removed rather than
edited to exclude the relevant portion. In the process of fixing this, mremap
has been extended to support MREMAP_FIXED and the extending of mappings (as
well as support for MREMAP_MAYMOVE not being selected, which will cause mmap
to return with -ENOMEM if a mapping could not be extended). The mmap runtime
test includes checks for these additional features.