Page table recurse_ptes(): avoid use of pte_is_block_mapping()
In RISC-V, the kernel does not have a way to distinguish a block
mapping from a page mapping without knowing the PTE level;
therefore, the pte_is_block_mapping() function implementation for
RISC-V is being removed, and the recurse_ptes() function, which was
the only caller of pte_is_block_mapping(), now calls
pte_is_mapping() instead.
This does not change the behavior of recurse_ptes() on x86 and ARM,
and fixes an issue on RISC-V that was causing recurse_ptes() to
wrongly recurse block mappings as if they were next-level PTEs,
therefore calling the PTE handler closure with invalid PTEs.