Make deferred tcp processing optional
Instead of queueing all tcp work, use mutex_try_lock and then process
the tcp work immediately if the mutex can be locked. This requires
some deferred tcp work handlers to know if the work is actually deferred
to determine whether or not to grab locks. This change allows for better
performance in uniprocessor machines.

The lwip service thunk now also holds the lwip lock as long as there
is work to do.