HTTP parser: add support for parsing split HTTP messages
When an HTTP message being received is split in multiple TCP
packets, the `http_recv` closure is invoked multiple times (once
for each packet). The existing code was invoking the value handler
at the first packet (with an incomplete HTTP message) and then
resetting the parser state (which would cause an incorrect parsing
of subsequent packets).
This commit changes the parser implementation so that the parser
state is not reset at each packet, and the value handler is invoked
only when a complete HTTP message has been received.