Support mounting volumes read-only
Support mounting tfs volumes read-only by checking for a string ":ro" appended
to the path of a volume in the mounts tuple. While tfs already supported a flag
for a read-only filesystem, support had to be added for checking the flag and
returning an appropriate error that could be translated into EROFS for syscalls.

The root volume can also be mounted readonly by specifying
"readonly_rootfs" in the manifest. This setting is not applied until
just before program load so that any nanos initialization that requires
touching the root filesystem can do so.

Additionally, the filesystem_creat_unnamed function needed to change to support
returning errors.