x86 Firecracker boot: fix initial mapping when RAM is > 1GB
Setting up the initial mapping requires the memory in the initial
pages region to be mapped. When booting the kernel via Firecracker,
this requirement was satisfied only if the VM was configured with a
RAM size up to 1 GB, because Firecracker identity-maps the first GB
of physical memory before starting the kernel, but was not
satisified if the initial pages region was allocated from physical
memory beyond 1GB.
This commit fixes the above issue by ensuring the initial pages
region is mapped before doing the initial mapping. In addition, the
pages being used for the temporary page tables are now reclaimed
after the initial mapping is in place, by using the newly defined
region_resize() function. The initialization code in
platform/pc/service.c has been refactored by defining a few new
functions, which will be reused when adding support for PVH boot.
The map_setup_2mbpages() has been amended so that it can be safely
called more than once, without corrupting any existing mappings by
overwriting higher-level page tables.