x86: add support for microvm Qemu machine type
With this change, the kernel supports running on the microvm Qemu
machine type, booting via the HVM direct boot protocol and using
virtIO-MMIO disk and network devices.
The memory location and interrupt number of virtIO-MMIO devices is
retrieved via ACPI by calling the newly defined function
virtio_mmio_enum_devs(). The code in virtio_mmio.c has been
refactored so that it can be used to enumerate devices via either
the kernel command line (used by Firecracker) or ACPI tables (used
by microvm). In addition, it now supports non-page-aligned memory
base values, and resets the devices being enumerated (which is
needed in case a device has been initialized by code running before
the kernel, such as the BIOS).
The MICROVM Makefile variable can be used to run a unikernel image
with microvm, as in `make run MICROVM=1`. The 'ioapic2=off' Qemu
machine option is necessary in recent Qemu versions in order to
prevent virtIO-MMIO interrupts from being routed to a secondary IO
APIC, which is not supported in the kernel. Since microvm does not
have a PCI bus, devices that sit on this bus (such as
virtio-rng-pci) are not instantiated when microvm is selected.
Running an image with microvm results in faster boot times and less
memory consumption compared to using the default q35 Qemu machine
type.